feat(ContactForm): Decoupling of CORE form submission logic, whilst building modular logic throughout the dynamic rendering of UI elements
