--- public/assets/js/contact.js
@@ -1,11 +1,0 @@
-
-class Msg{
-
-constructor(){
-
-}
-
-}
-
-new Msg();
-
