+++ public/assets/js/logic/validator.js
@@ -1,0 +1,35 @@
+
+
+/* Designation of the CLASS used to validate our user input. */
+export class Validator {
+
+    static REQUIRED = "REQUIRED";
+    static MIN_LENGTH = "MIN_LENGTH";
+    static NUMBER = "NUMBER";
+    static MAX_LENGTH = "MAX_LENGTH";
+    static DROP_BOX = "DROP_BOX";
+    static EMAIL = "EMAIL";
+
+static validate(value, flag, validatorValue){
+    if(flag === this.REQUIRED){
+        return value.trim().length > 0;
+    }
+    if(flag === this.MIN_LENGTH){
+        return value.trim().length > validatorValue;
+    }
+    if(flag === this.MAX_LENGTH){
+        return value.length < validatorValue;
+    }
+    if(flag === this.NUMBER){
+        return isNaN(value) ;
+    }
+    if(flag === this.DROP_BOX){
+        return value > 0;
+    }
+    if(flag === this.EMAIL){
+        return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value);
+    }
+}
+
+
+}
