+++ public/index.html
@@ -36,7 +36,7 @@
             <link rel="stylesheet" href="assets/css/styles.css"> 
             <!-- JS Functionality -->
             <!-- <script src="./assets/js/main.js" async></script> -->
-            <script src="./assets/js/app.js" async></script>
+            <script type="module" src="./assets/js/app.js" async></script>
 
 
         </head>
